pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Shovik"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "joblib>=1.4.2",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "replit>=4.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
]
